apps/crawler/requirements.txt
uvicorn
pytest-playwright
playwright 
pydantic-settings
newspaper3k
"lxml[html_clean]"
feedparser
aiohttp
googlenewsdecoder
asyncpg
schedule
watchdog
httpx